asim: extract history into its own package
Previously, the history struct (capturing store metrics at each tick and
simulator’s current state) resided within the asim package.

Another PR(#109316) is introducing assertion registration which enables
scheduled assertion checks to run as delayed events. Since these assertion
events use simulation’s history to conduct assertion checks, the event package
would have to depend on the asim package. However, the asim package would also
have to depend on the event package’s event executor for event ticking, creating
a circular dependency.

To address this issue, this patch moves the history component out of the asim
package to its own package.

Note that this commit does not change any existing behavior, and the main
purpose is to make future commits cleaner.

See also: #109316
Part of: #106192
Release note: none
