+++ pkg/kv/kvserver/asim/history/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "history",
+    srcs = ["history.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim/history",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/kv/kvserver/asim/metrics",
+        "//pkg/kv/kvserver/asim/state",
+    ],
+)
